+++ packages/flutter_user/lib/src/flutter_user_navigator_userstory.dart
@@ -2,7 +2,6 @@ import "dart:async";
 
 import "package:flutter/material.dart";
 import "package:flutter_user/flutter_user.dart";
-import "package:user_repository_interface/user_repository_interface.dart";
 
 class FlutterUserNavigatorUserstory extends StatefulWidget {
   const FlutterUserNavigatorUserstory({
@@ -12,6 +11,7 @@ class FlutterUserNavigatorUserstory extends StatefulWidget {
     this.options,
     this.forgotPasswordTranslations,
     this.registrationOptions,
+    this.forgotPasswordOptions = const ForgotPasswordOptions(),
     super.key,
   });
 
@@ -23,6 +23,7 @@ class FlutterUserNavigatorUserstory extends StatefulWidget {
   final VoidCallback? afterRegistration;
   final ForgotPasswordTranslations? forgotPasswordTranslations;
   final RegistrationOptions? registrationOptions;
+  final ForgotPasswordOptions forgotPasswordOptions;
 
   @override
   State<FlutterUserNavigatorUserstory> createState() =>
@@ -102,10 +103,10 @@ class _FlutterUserNavigatorUserstoryState
         );
 
         if (!loginResponse.loginSuccessful) {
-          if (context.mounted) {
-            // ignore: use_build_context_synchronously
-            await errorScaffoldMessenger(context, loginResponse);
-          }
+          if (!mounted) return;
+          Navigator.of(context).pop();
+          if (!context.mounted) return;
+          await errorScaffoldMessenger(context, loginResponse);
           return;
         }
         await options!.afterLogin?.call();
@@ -162,6 +163,7 @@ class _FlutterUserNavigatorUserstoryState
       title: title,
       description: description,
       loginOptions: options!.loginOptions,
+      forgotPasswordOptions: widget.forgotPasswordOptions,
       onRequestForgotPassword: (email) async {
         if (options!.onRequestForgotPassword != null) {
           await options!.onRequestForgotPassword!(email);
@@ -222,7 +224,7 @@ class _FlutterUserNavigatorUserstoryState
             return 1;
           }
           if (isEmailError) {
-            return 2;
+            return 0;
           }
 
           return null;